added timer to eos deactivate// still need to set up peramiters for it
eos deactivate timer is working, default 5 minutes, need it to work with description.ext. as of now it is not allowing to respawn if the parameters are in the descript file, no idea why, have a feeling im doing something wrong with setting up parameters . need to get to sleep. almost 4